Got like_count comment_count, user and caption consistently.
Now we need to learn how to :
Click on an God damn element.
Brute force solution: get element coordinates then click that xy point. This is stupid and not recommended but im losing my mind why is it so hard to consistently click an element it works perfectly in the devtools console FUCK bruh
We also need to share a post, leave a comment, visit profile then go back but I think these are easy. theyll be done next commit.